+++ src/main/java/com/daeunchung/wmslive/product/domain/ProductRepository.java
@@ -1,11 +1,11 @@
-package com.daeunchung.wmslive.product.feature;
+package com.daeunchung.wmslive.product.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-class ProductRepository {
+public class ProductRepository {
     private final Map<Long, Product> products = new HashMap<>();
     private Long nextId = 1L;
